Fbx Animation Export 09
Smoother pages (if need be I can smooth them again). UV maps are appropriate for the page proportions. I’m giving you two versions, one with traditionally laid out UV’s (04_BookBakev09_UVLayout) and the other with them overlapping (04_BookBakev09_UVoverlap). 

I used Blend Shapes for a more suitable animation. The animation is seamless. The following is a breakdown of the frame ranges and their animation:

1-13: Open book
13-33: Flip Left
33-50: Flip Right
50-62: Close book

Frames that overlap are as follows: 13, 33, 50.

The following files have been exported as an FBX.